pyproject.toml
[project]
name = "europe-or-not"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.10.8",
    "openai>=2.16.0",
    "pillow>=12.1.0",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "tqdm>=4.67.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/europe_or_not"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
